set aside functions no longer useful
